fix : es6 grammar
- es6문법으로 수정

+++ todoList.js
@@ -4,56 +4,68 @@ function todoList() {
     const inputTodo = document.getElementById('inputTodo')
     const clearBtn = document.getElementById('clearBtn')
     const ulTodo = document.getElementById('todoList')
-    const middleSection = document.getElementById('middleSection')
-    const lowerSection = document.getElementById('lowerSection')
+    const listBox = document.getElementById('app')
 
-    // event listener
-    addBtn.addEventListener('click', clickAddBtn)
-    clearBtn.addEventListener('click', clearTodoList)
-    ulTodo.addEventListener('click', function(e) { clickTodoBtn(e) })
+    addBtn.addEventListener('click', e => clickAddBtn(e))
+    clearBtn.addEventListener('click', () => clearTodoList())
+    ulTodo.addEventListener('click', e => clickTodoBtn(e))
 
+    // 로컬스토리지 불러오기 & 저장하기
+    const getLocalStorage = () => JSON.parse(localStorage.getItem('todo')) || []
+    const setLocalStorage = (list) => localStorage.setItem('todo', JSON.stringify(list))
 
-    function getLocalStorage() {
-        let savedArr = JSON.parse(localStorage.getItem('todo')) || []
-        return savedArr
-    }
-
-    function setLocalStorage(list) {
-        localStorage.setItem('todo', JSON.stringify(list))
-    }
-
-    function clickAddBtn(e) {
+    // 클릭 이벤트 : 리스트 추가 실행
+    const clickAddBtn = (e) => {
         e.preventDefault()
-        if (inputTodo.value === '') {
-            return
-        }
-        addTodo()
+        if (inputTodo.value !== '') addTodo()
     }
 
-    function addTodo() {
+    // 리스트 추가
+    const addTodo = () => {
         const newTodo = { text: inputTodo.value, checked: false }
-        inputTodo.value = ''
         const newArr = [...getLocalStorage(), newTodo]
+        inputTodo.value = ''
         showTodoList(newArr)
-
     }
 
-    function clickTodoBtn(e) {
-        if (e.target.id === 'checkBtn') {
-            checkTodo(e)
-        } else if (e.target.id === 'deleteBtn') {
-            deleteTodo(e)
-        }
+    // 리스트 렌더링 & 로컬스토리지 저장 실행
+    const showTodoList = (list) => {
+        setLocalStorage(list)
+        renderTodoList(list)
+        sectionInactive(list)
     }
 
-    function deleteTodo(e) {
+    // 리스트 렌더링
+    const renderTodoList = (list) => {
+        ulTodo.innerHTML = list.map((todo, i) => `<li class="todo ${todo.checked===true ? "checked" : ""}" data-num="${i}">
+            <button id="checkBtn"><i class="fa-solid fa-check"></i></button>
+            <span>${todo.text}</span>
+            <button id="deleteBtn"><i class="fa-solid fa-xmark"></i></button
+        </li>`).join('')
+    }
+
+    // 리스트가 없는 경우 secton에 display : none 처리
+    const sectionInactive = (list) => (list.length === 0) ? listBox.classList.add('inactive') : listBox.classList.remove('inactive')
+
+    // 초기 화면 로딩
+    const init = () => showTodoList(getLocalStorage())
+    init()
+
+    // 클릭 이벤트 : 해당 투두 체크 또는 삭제
+    const clickTodoBtn = (e) => {
+        if (e.target.id === 'checkBtn') checkTodo(e)
+        if (e.target.id === 'deleteBtn') deleteTodo(e)
+    }
+
+    // 해당 투두 삭제
+    const deleteTodo = (e) => {
         const deleteNum = e.target.closest('li').dataset.num
-        const savedArr = getLocalStorage()
-        const newArr = savedArr.filter((todo, i) => i !== Number(deleteNum))
+        const newArr = getLocalStorage().filter((todo, i) => i !== Number(deleteNum))
         showTodoList(newArr)
     }
 
-    function checkTodo(e) {
+    // 해당 투두 체크 표시
+    const checkTodo = (e) => {
         const checkNum = e.target.closest('li').dataset.num
         const savedArr = getLocalStorage()
         const newArr = [...savedArr]
@@ -61,40 +73,9 @@ function todoList() {
         showTodoList(newArr)
     }
 
-    function clearTodoList() {
-        showTodoList([])
-    }
+    // 클릭이벤트 : 투두리스트 전체 삭제 실행
+    const clearTodoList = () => showTodoList([])
 
-    function showTodoList(list) {
-        setLocalStorage(list)
-        renderTodoList(list)
-    }
-
-    function renderTodoList(list) {
-        sectionInactive(list)
-        const liTodo = list.map((todo, i) => `<li class="todo ${todo.checked===true ? "checked" : ""}" data-num="${i}">
-            <button id="checkBtn"><i class="fa-solid fa-check"></i></button>
-            <span>${todo.text}</span>
-            <button id="deleteBtn"><i class="fa-solid fa-xmark"></i></button
-        </li>`).join('')
-        ulTodo.innerHTML = liTodo
-    }
-
-    function sectionInactive(list) {
-        if (list.length === 0) {
-            middleSection.classList.add('inactive')
-            lowerSection.classList.add('inactive')
-        } else {
-            middleSection.classList.remove('inactive')
-            lowerSection.classList.remove('inactive')
-        }
-    }
-
-    function init() {
-        const initArr = getLocalStorage()
-        renderTodoList(initArr)
-    }
-    init()
 }
 
 window.addEventListener('DOMContentLoaded', todoList)
